% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookdown.r
\name{adj_yaml_non_work}
\alias{adj_yaml_non_work}
\title{Include or exclude non-work Rmds from _bookdown.yml}
\usage{
adj_yaml_non_work(inc_non_work = FALSE)
}
\description{
Include or exclude non-work Rmds from _bookdown.yml
}
