% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.r
\name{assert}
\alias{assert}
\title{Check class of an object}
\usage{
assert(x, y)
}
\arguments{
\item{x}{object. Object whose class is assessed.}

\item{y}{character vector. Expected classes of \code{x}.}
}
\description{
Meant to not be run interactively,
but added to code without
adding depenencies, and so is not
exported.
}
