% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_data_frame.r
\name{print.data.frame}
\alias{print.data.frame}
\title{Print data frame like head}
\usage{
\method{print}{data.frame}(df)
}
\arguments{
\item{df}{dataframe.}
}
\description{
Print data frame like head
}
