% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detach.r
\name{.detach_pkg}
\alias{.detach_pkg}
\title{Detach a package}
\usage{
.detach_pkg(x)
}
\arguments{
\item{x}{character vector.
Names of packages to detach.}
}
\description{
Wrapper function around library and detach
to detach packages.
}
