% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_dir.R
\name{.od}
\alias{.od}
\title{Open a dialogue box at selected folder}
\usage{
.od(path = here::here())
}
\arguments{
\item{path}{character. The initial working directory,
from which the file dialog should begin browsing.
Defaults to \code{here::here()}.}
}
\value{
\code{invisible(<selected_file>/NULL)} (\code{NULL} if no file were
  selected).
}
\description{
Useful for easily opening a window in
which you can see all the files in a folder and open them using
\code{Open with...}.
}
\examples{
od()
}
