% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_vsc_setting-r_version.r
\name{.set_r_version}
\alias{.set_r_version}
\title{Set R version for build and console}
\usage{
.set_r_version(
  r_v = "",
  nm = c("r.rpath.windows", "r.rterm.option"),
  path_to_settings = here::here(".vscode/settings.json"),
  r_dir = "C:\\\\Program Files\\\\R"
)
}
\arguments{
\item{r_v}{character. R version.
If "3" or "4", then set to "3.6.3" or "4.1.2", respectively.
Otherwise must specify full version.}

\item{nm}{character R\.set_vsc_setting-r_version.rvector.
VSC options to set R version for.
At present, must be options starting either
with \code{"r.rpath"} or \code{"r.term"}.
Default is \code{c("r.rpath.windows", "r.rterm.option")}.}

\item{r_dir}{character. Path where "R-<version>/bin/R.exe" is.}
}
\value{
\code{invisible(TRUE)}. Side effect is that
JSON file with settings is updated.
}
\description{
Set R version in VSC settings.
}
