% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_vsc_setting.r
\name{.set_vsc_setting_rmd}
\alias{.set_vsc_setting_rmd}
\title{Set VSC Rmd settings}
\usage{
.set_vsc_setting_rmd(
  editor.wordWrap = "on",
  editor.quickSuggestions = FALSE,
  editor.formatOnType = FALSE,
  path_to_settings = here::here(".vscode/settings.json")
)
}
\description{
Set typical VSC settings with respect to RMarkdown documents.
}
