% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_r_version.r
\name{.srv}
\alias{.srv}
\title{Set R version in VSC workspace settings}
\usage{
.srv(
  r_v = "",
  r_dir = "C:\\\\Program Files\\\\R",
  ws_settings = c("r.rpath.windows", "r.rterm.option")
)
}
\arguments{
\item{r_v}{character. R version.
If "3" or "4", then set to "3.6.3" or "4.1.2", respectively.
Otherwise must specify full version.}

\item{r_dir}{character. Path where "R-<version>/bin/R.exe" is.}

\item{ws_settings}{character. VS Code workspace settings to set.}
}
\value{
\code{invisible(TRUE)}.
}
\description{
Set R version in VSC workspace settings.
}
