% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_vsc_setting.r
\name{.unset_vsc_setting}
\alias{.unset_vsc_setting}
\title{Unset a VSC setting}
\usage{
.unset_vsc_setting(nm, path_to_settings = here::here(".vscode/settings.json"))
}
\arguments{
\item{nm}{character vector.
Names of settings to remove.}
}
\description{
Remove a setting entirely from the VSC JSON
settings file.
}
