% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_r_version.r
\name{.urv}
\alias{.urv}
\title{Remove R version specification from VSC workspace}
\usage{
.urv(ws_settings = c("r.rpath.windows", "r.rterm.option"))
}
\arguments{
\item{ws_settings}{character vector. names of VSC workspace
settings that should be removed to unset R version.}
}
\value{
\code{invisible(TRUE)}.
}
\description{
Unset R version from being fixed by VSC workspace settings.
This function just removes the JSON elements
specified by \code{ws_settings} if they're there, and does nothing otherwise.
}
