% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookdown.r
\name{kb}
\alias{kb}
\title{Knit book}
\usage{
kb(input = NULL, exact = FALSE, inc_ext = FALSE)
}
\arguments{
\item{input}{character vector. If not \code{NULL}, then files
that match to elements in here
are knitted.
If \code{NULL}, thena all files are knitted.
Default is \code{NULL}.}

\item{exact}{logical.
If \code{FALSE}, then partial matches for \code{input} are accepted.
Default is \code{FALSE}.}

\item{inc_ext}{logical.
If \code{TRUE}, then files in \code{ext/} are included.
Default is \code{FALSE}.}
}
\description{
Knit book
}
