% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od.r, R/open_dir.R
\name{od}
\alias{od}
\title{Open File Explorer at a particular directory}
\usage{
od(path = here::here())

od(path = here::here())
}
\arguments{
\item{path}{character. The initial working directory,
from which the file dialog should begin browsing.
Defaults to the current RStudio project directory.}
}
\value{
\code{invisible(<selected_file>/NULL)} (\code{NULL} if no file were
  selected).
}
\description{
Useful for easily opening a window in
which you can see all the files in a folder and open them using
\code{Open with...}.
}
\details{
Only works in Windows, at it calls PowerShell (unless
PowerShell can work on multiple OS's).
}
\examples{
od()
}
