% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_radian_with_new_r_version.r
\name{r_cv}
\alias{r_cv}
\title{Change R terminal version between 3.6.3 and 4.1.0

Does not seem to work properly - seems to still use old libraries}
\usage{
r_cv(version = NULL)
}
\description{
Change R terminal version between 3.6.3 and 4.1.0

Does not seem to work properly - seems to still use old libraries
}
