% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_radian_with_new_r_version.r
\name{r_ov}
\alias{r_ov}
\title{Open a specified R Version in terminal}
\usage{
r_ov(version)
}
\description{
Open a specified R Version in terminal
}
\details{
Only seems to work when opening current version.
Therefore only r_r is exported.
}
